--- MusicApp/app/src/main/java/com/hoaiphong/composeui/ui/screen/login/LoginButton.kt
@@ -1,33 +1,0 @@
-package com.hoaiphong.composeui.ui.screen.login
-
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun LoginButton(
-    text: String = "Log in",
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit = {}
-) {
-    Button(
-        onClick = onClick,
-        modifier = modifier
-            .fillMaxWidth()
-            .height(50.dp),
-        colors = ButtonDefaults.buttonColors(
-            containerColor = Color(0xFF00BCD4),
-            contentColor = Color.White
-        ),
-        shape = CircleShape
-    ) {
-        Text(text = text)
-    }
-}
