--- MusicApp/app/src/main/java/com/hoaiphong/composeui/ui/screen/login/PasswordInput.kt
@@ -1,81 +1,0 @@
-package com.hoaiphong.composeui.ui.screen.login
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.OutlinedTextFieldDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.unit.dp
-import com.hoaiphong.composeui.R
-
-@Composable
-fun PasswordInput(
-    password: String = "",
-    onPasswordChange: (String) -> Unit = {},
-    labelText: String = "Password"
-) {
-    var passwordVisible by remember { mutableStateOf(false) }
-
-    OutlinedTextField(
-        value = password,
-        onValueChange = onPasswordChange,
-        label = { Text(labelText, color = Color.Gray) },
-        leadingIcon = {
-            Icon(
-                imageVector = Icons.Default.Lock,
-                contentDescription = "Password Icon",
-                tint = Color.Gray
-            )
-        },
-        trailingIcon = {
-            val iconResId = if (passwordVisible)
-                R.drawable.ic_visible
-            else
-                R.drawable.ic_enable
-
-            IconButton(onClick = { passwordVisible = !passwordVisible }) {
-                Icon(
-                    painter = painterResource(id = iconResId),
-                    contentDescription = if (passwordVisible) "Hide password" else "Show password",
-                    tint = Color.Gray
-                )
-            }
-        },
-        visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
-        colors = OutlinedTextFieldDefaults.colors(
-            focusedBorderColor = Color.Gray,
-            unfocusedBorderColor = Color.Gray,
-            cursorColor = Color.Cyan,
-            focusedLeadingIconColor = Color.Gray,
-            unfocusedLeadingIconColor = Color.Gray,
-            focusedTrailingIconColor = Color.Gray,
-            unfocusedTrailingIconColor = Color.Gray,
-            unfocusedTextColor = Color.White,
-            focusedTextColor = Color.White,
-            focusedLabelColor = Color.Gray,
-            unfocusedLabelColor = Color.Gray
-        ),
-        shape = RoundedCornerShape(12.dp),
-        singleLine = true,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(56.dp)
-            .background(Color(0xFF1A1A1A), shape = RoundedCornerShape(12.dp))
-    )
-}
