--- MusicApp/app/src/main/java/com/hoaiphong/composeui/ui/screen/login/TextInput.kt
@@ -1,60 +1,0 @@
-package com.hoaiphong.composeui.ui.screen.login
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material3.Icon
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.OutlinedTextFieldDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun TextInput(
-    value: String = "",
-    onValueChange: (String) -> Unit = {},
-    leadingIcon: ImageVector = Icons.Default.Person,
-    placeholderText: String = "Username",
-    visualTransformation: VisualTransformation = VisualTransformation.None
-) {
-    OutlinedTextField(
-        value = value,
-        onValueChange = onValueChange,
-        placeholder = { Text(placeholderText, color = Color.Gray) },
-        leadingIcon = {
-            Icon(
-                imageVector = leadingIcon,
-                contentDescription = null,
-                tint = Color.Gray
-            )
-        },
-        visualTransformation = visualTransformation,
-        colors = OutlinedTextFieldDefaults.colors(
-            focusedBorderColor = Color.Gray,
-            unfocusedBorderColor = Color.Gray,
-            cursorColor = Color.Cyan,
-            focusedLeadingIconColor = Color.Gray,
-            unfocusedLeadingIconColor = Color.Gray,
-            focusedPlaceholderColor = Color.Gray,
-            unfocusedPlaceholderColor = Color.Gray,
-            unfocusedTextColor = Color.White,
-            focusedTextColor = Color.White,
-            disabledTextColor = Color.LightGray,
-            disabledBorderColor = Color.DarkGray
-        ),
-        shape = RoundedCornerShape(12.dp),
-        singleLine = true,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(56.dp)
-            .background(Color(0xFF1A1A1A), shape = RoundedCornerShape(12.dp))
-    )
-}
